# src/openapi.yaml
openapi: 3.0.3
info:
  title: InvoiceSmart API
  version: 1.0.0
  description: |
    REST API powering the InvoiceSmart iOS app (and any future web client).
    Multi-tenant: every authenticated request is scoped to the JWT bearer's
    `user_id`. Cross-tenant reads return 404 (not 403) to avoid leaking
    existence.

    **Auth**: `POST /api/auth/login` returns a JWT, valid 7 days. Send as
    `Authorization: Bearer <jwt>` on every protected route. The `/health`
    and `/api/health` routes are unauthenticated and report DB
    connectivity for L7 probes.

    **Money** is `DECIMAL(12,2)` server-side and returned as a string
    (e.g. `"1500.00"`) — clients should `parseFloat` before arithmetic.
    Rates (`tax_rate`, `discount_rate`, `cis_rate`, `retention_rate`) are
    percentages, so `20` means 20% — never `0.20`.
servers:
  - url: https://api.invoicesmart.cortexbuildpro.com
    description: Production
  - url: http://127.0.0.1:3008
    description: Local dev / docker
...
security:
  - bearerAuth: []
paths:
...
  /api/transactions:
    get:
      tags: [transactions]
      summary: List transactions (filterable by type + date range)
      parameters:
        - { in: query, name: type, schema: { type: string, enum: [expense, payment] } }
        - { in: query, name: startDate, schema: { type: string, format: date } }
        - { in: query, name: endDate, schema: { type: string, format: date } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TransactionList' } } } }
    post:
      tags: [transactions]
      summary: Record a standalone transaction (typically an expense)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TransactionInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } } }
  /api/transactions/invoice/{invoiceId}:
    parameters:
      - { in: path, name: invoiceId, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [transactions]
      summary: List transactions linked to one invoice
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Transaction' } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/transactions/{id}:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    delete:
      tags: [transactions]
      summary: Delete a transaction
      responses:
        '204': { description: Deleted }
  /api/tax-rules:
    get:
      tags: [tax-rules]
      summary: List tax rules for the current user
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRuleList' } } } } }
    post:
      tags: [tax-rules]
      summary: Create a tax rule
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRuleInput' } } } }
      responses: { '201': { description: Created } }
  /api/tax-rules/{id}:
    parameters: [{ $ref: '#/components/parameters/UuidPath' }]
    put:
      tags: [tax-rules]
      summary: Update a tax rule
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRuleInput' } } } }
      responses: { '200': { description: OK } }
    delete:
      tags: [tax-rules]
      summary: Delete a tax rule
      responses: { '204': { description: Deleted } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: Must be a valid v4-style UUID; malformed values return 400.
  responses:
...
    NotFound:
      description: Entity not found (or belongs to a different tenant)
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }
  schemas:
...
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        invoice_id: { type: string, format: uuid, nullable: true }
        type: { type: string, enum: [expense, payment] }
        amount: { type: string }
        transaction_date: { type: string, format: date }
        category: { type: string, nullable: true }
        description: { type: string, nullable: true }
        payment_method: { type: string, nullable: true }
        reference: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    TransactionInput:
      type: object
      required: [type, amount]
      properties:
        invoice_id: { type: string, format: uuid }
        type: { type: string, enum: [expense, payment] }
        amount: { type: number, minimum: 0.01 }
        transaction_date: { type: string, format: date }
        category: { type: string }
        description: { type: string }
        reference: { type: string }
    TransactionList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Transaction' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    TaxRule:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        rate: { type: string }
        type: { type: string, example: vat }
        country: { type: string, example: GB }
        is_default: { type: boolean }
    TaxRuleInput:
      type: object
      required: [name, rate]
      properties:
        name: { type: string }
        rate: { type: number, minimum: 0 }
        type: { type: string, default: vat }
        country: { type: string }
        is_default: { type: boolean }
    TaxRuleList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/TaxRule' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }